+++ extensions/httpclient/src/test/groovy/com/stormpath/sdk/client/DirectoryIT.groovy
@@ -16,7 +16,13 @@
 package com.stormpath.sdk.client
 
 import com.stormpath.sdk.account.Account
+import com.stormpath.sdk.account.AccountStatus
 import com.stormpath.sdk.account.Accounts
+import com.stormpath.sdk.account.EmailVerificationStatus
+import com.stormpath.sdk.account.PasswordFormat
+import com.stormpath.sdk.application.Application
+import com.stormpath.sdk.authc.AuthenticationResult
+import com.stormpath.sdk.authc.UsernamePasswordRequests
 import com.stormpath.sdk.directory.AccountCreationPolicy
 import com.stormpath.sdk.directory.Directories
 import com.stormpath.sdk.directory.Directory
@@ -1259,6 +1265,79 @@ class DirectoryIT extends ClientIT {
 
         assertNull(account.getGivenName())
         assertNull(account.getSurname())
+    }
+
+    /**
+     * @since 1.5.3
+     */
+    @Test
+    void testBCryptPasswordForAccountCreatedViaDirectory() {
+        Application application = createTempApp()
+        Directory directory = client.instantiate(Directory)
+        directory.setName(uniquify("JSDK.DirectoryIT.testGetOrganizations"))
+        directory = client.createDirectory(directory)
+        assertNotNull directory.href
+        deleteOnTeardown(directory)
+
+        application.addAccountStore(directory)
+
+        Account account = client.instantiate(Account.class)
+        String username = "deleteme" + UUID.randomUUID()
+        account.setEmail(username + "@mailinator.com")
+                .setUsername(username)
+                .setGivenName(username)
+                .setSurname(username)
+                .setPassword('$2a$12$PLmHvcbeliNkjIVFyVXg5O.LQfyYtHFm.1vAtj3l7itIP.z0V6OX2') //hash for Pa$sw0rd123
+
+        directory.createAccount(Accounts.newCreateRequestFor(account).setPasswordFormat(PasswordFormat.MCF).build())
+
+        AuthenticationResult authenticationResult = application.authenticateAccount(
+                UsernamePasswordRequests.builder()
+                        .setUsernameOrEmail(username)
+                        .setPassword('Pa$sw0rd123')
+                        .withResponseOptions(UsernamePasswordRequests.options().withAccount())
+                        .build())
+
+        assertEquals(authenticationResult.getAccount().getUsername(), username)
+    }
+
+    /**
+     * @since 1.5.4
+     */
+    @Test
+    void testBCryptPasswordForAccountCreatedViaDirectoryWithRegistrationWorkflowOverride() {
+        Application application = createTempApp()
+        Directory directory = client.instantiate(Directory)
+        directory.setName(uniquify("JSDK.DirectoryIT.testBCryptPasswordForAccountCreatedViaDirectoryWithRegistrationWorkflowOverride"))
+        directory = client.createDirectory(directory)
+        assertNotNull directory.href
+        deleteOnTeardown(directory)
+
+        application.addAccountStore(directory)
+
+        Account account = client.instantiate(Account.class)
+        String username = "deleteme" + UUID.randomUUID()
+        account.setEmail(username + "@mailinator.com")
+                .setUsername(username)
+                .setGivenName(username)
+                .setSurname(username)
+                .setStatus(AccountStatus.ENABLED)
+                .setEmailVerificationStatus(EmailVerificationStatus.VERIFIED)
+                .setPassword('$2a$12$PLmHvcbeliNkjIVFyVXg5O.LQfyYtHFm.1vAtj3l7itIP.z0V6OX2') //hash for Pa$sw0rd123
+
+        directory.createAccount(Accounts.newCreateRequestFor(account)
+                                        .setPasswordFormat(PasswordFormat.MCF)
+                                        .setRegistrationWorkflowEnabled(false)
+                                        .build())
+
+        AuthenticationResult authenticationResult = application.authenticateAccount(
+                UsernamePasswordRequests.builder()
+                        .setUsernameOrEmail(username)
+                        .setPassword('Pa$sw0rd123')
+                        .withResponseOptions(UsernamePasswordRequests.options().withAccount())
+                        .build())
+
+        assertEquals(authenticationResult.getAccount().getUsername(), username)
     }
 
 }